feat: save intermediate trained models during challenge + demo (#108)
* saves algorithm
* modified lds intermediate plot for demo
* added pickling + demo, changes to intermediate plotting
