+++ src/dynadojo/systems/lds.py
@@ -4,9 +4,11 @@ Linear Dynamical Systems
 """
 import numpy as np
 import scipy as sp
+import matplotlib.pyplot as plt
 
 from .utils.simple import SimpleSystem
 from ..utils.lds import plot
+
 
 
 class LDSystem(SimpleSystem):
@@ -123,8 +125,9 @@ class LDSystem(SimpleSystem):
     def save_plotted_trajectories( self, 
             y_true:np.ndarray, 
             y_pred: np.ndarray,
-            filepath: str,
+            filepath: str = "LDSystem_plotted_trajectories.pdf",
             tag: str = "", 
+            savefig: bool = True
         ):
         """
         Plots the trajectories of the system and the predicted trajectories.
@@ -141,5 +144,9 @@ class LDSystem(SimpleSystem):
                        labels=["true", "pred"], 
                        max_lines=10,
                        title=f"LDS l={self.latent_dim}, e={self.embed_dim} - {tag}")
-        fig.savefig(filepath, bbox_inches='tight', dpi=300, transparent=True, format='pdf')
-        return fig, ax
+        if savefig:
+            fig.savefig(filepath, bbox_inches='tight', dpi=300, transparent=True, format='pdf')
+            plt.close(fig)
+            return None, None
+        else:
+            return fig, ax
