+++ src/dynadojo/systems/lv/prey_predator.py
@@ -3,6 +3,7 @@ Prey-Predator Lotka Volterra
 """
 import numpy as np
 from scipy.integrate import solve_ivp
+import matplotlib.pyplot as plt
 
 from ...abstractions import AbstractSystem
 from ...utils.lv import plot
@@ -245,8 +246,9 @@ class PreyPredatorSystem(AbstractSystem):
     def save_plotted_trajectories( self, 
             y_true:np.ndarray, 
             y_pred: np.ndarray,
-            filepath: str,
+            filepath: str = "PreyPredatorSystem_plotted_trajectories.pdf",
             tag: str = "", 
+            savefig: bool = True
         ):
         """
         Plots the trajectories of the system and the predicted trajectories.
@@ -263,5 +265,9 @@ class PreyPredatorSystem(AbstractSystem):
                        labels=["true", "pred"], 
                        max_lines=10,
                        title=f"LV (C) l={self.latent_dim}, e={self._embed_dim} - {tag}")
-        fig.savefig(filepath, bbox_inches='tight', dpi=300, transparent=True, format='pdf')
-        return fig, ax
+        if savefig:
+            fig.savefig(filepath, bbox_inches='tight', dpi=300, transparent=True, format='pdf')
+            plt.close(fig)
+            return None, None
+        else:
+            return fig, ax
